requirements.txt
requests
web3
opencv-python
numpy
python-dotenv
